load_tests/requirements.txt
certifi==2018.8.13
chardet==3.0.4
click==6.7
Flask==1.0.2
gevent==1.3.6
greenlet==0.4.14
idna==2.7
itsdangerous==0.24
Jinja2==2.10
locustio==0.8.1
MarkupSafe==1.0
msgpack-python==0.5.6
pyzmq==17.1.2
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
